Updated with glitch

+++ skills/file_upload.js
@@ -5,7 +5,7 @@ const fs = require('fs');
 
 module.exports = function(controller) {
   
-  controller.file_upload = function(bot, message, cb) {
+  controller.fileUpload = function(bot, message, cb) {
     var destination_path = 'tmp/uploaded/';
 
     // the url to the file is in url_private. there are other fields containing image thumbnails as appropriate
@@ -18,8 +18,10 @@ module.exports = function(controller) {
           Authorization: 'Bearer ' + bot.config.bot.token, // Authorization header with bot's access token
         }
     };
+    
+    var title = message.user + "_" + message.ts;
 
-    var filePath = destination_path + message.file.title;
+    var filePath = destination_path + title;
 
     var stream = request(opts, function(err, res, body) {
         console.log('FILE RETRIEVE STATUS',res.statusCode);          
@@ -27,13 +29,15 @@ module.exports = function(controller) {
 
     stream.on("finish", function() {
       // When stream is finished, upload the file
-      cloudinary.v2.uploader.unsigned_upload(destination_path + message.file.title, "image_counter_bot", 
+      cloudinary.v2.uploader.unsigned_upload(filePath, "image_counter_bot", 
           { resource_type: "image", tags: [ 'user_' + message.user, 'team_' + message.team ] },
          function(err, result) {
         console.log(err, result);
         
-        // Remove the file from the temporary storage
-        fs.unlinkSync(filePath);
+        if (fs.existsSync(filePath)) {
+          // Remove the file from the temporary storage
+          fs.unlinkSync(filePath);
+        }
         
         // if we have a callback, run it
         if (cb) cb(result);
